show only the search overlay of the current track
this commit introduces the methods prepare() and cleanup()
of each Track (and the subclasses of tracks). They are being
called everytime, the current_track of TrackList changes.
The first use is to hide the searchoverlay of a track if it
is currently active and the track ceases to be the current_track.
if the track is selected as current_track again, the search
overlay will be restored.