use custom function to clear cache
the webkit default clear_cache function deletes the whole content
of the cachefolder. we want to keep our files like the stored
session, the cookies and the history DB. so i implemented a custom
function that deletes every file in the cachefolder that's filename
only consists of digits  we should add a feature
do deliberately delete the whole cache if the user wants to do so.