did some preparations for scroll-state feature
now basically everything is implemented that is
needed for sitenodes to remember where the user
has scrolled at the time the node was active.
the only feature is missing is a scroll_to-method
in the stable API of webkit2gtk. we are
waiting for them to move that method to stable